API: Preserve UUID in path table
Changes the AddPath API to add the UUID generated to the path object.
Previously this UUID was only usable for deleting routes, and
would be dropped when inserted into the table. That meant then if the
same route was passed to a table event watcher, the UUID was set to nil,
despite being made by the API.

With this change, the UUID is preserved and will be passed to callers of
ListPath, or when passed to an event watcher. To implement this an
additional UUID field has been added onto the table path object, and the
AddPath, api2Path functions modified to preserve them. The Marshaller
was also updated to return this value to end users as well.

Additionally the DeletePath call has been simplified when deleting paths
by UUID. Since each path now stores the UUID they are checked directly
rather than using a lookup map.